Added perm_processing.c (Almost done!), updated README.md & to-do-list.md files.
(Almost done - Potentially going to add more functions for checking permutations parameters and still need to tidy up the code a *little* bit more).
Also added the result.out file which I used for testing the output of perm_processing(.c) subprogram (Did not feature this update in the main comment because the file is rather very insignificant).